Rename stop command to stop-containers
